refactor: move context switching functions into compat
They exist to ensure compatibility between different userscript
engines and browsers, so they belong there.
